agents/pyproject.toml
[project]
name = "genai_boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "colorama>=0.4.6",
    "fastapi[standard]>=0.116.1",
    "jwt>=1.4.0",
    "openai>=1.97.0",
    "opentelemetry-api>=1.35.0",
    "opentelemetry-distro>=0.56b0",
    "opentelemetry-exporter-otlp>=1.35.0",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "greenlet>=3.2.3",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.4",
]

[tool.pytest.ini_options]
pythonpath = ["./app"]
addopts = "--cov=app"
asyncio_mode = "auto"
